// apps/api/src/property-leads/lead.enums.ts
export enum LeadOrigin {
  WEB = 'web',
  WHATSAPP = 'whatsapp',
  VALUATION = 'valuation',
  SELL_FORM = 'sell_form',
}

export enum LeadStatus {
  NEW = 'nuevo',
  CONTACTED = 'contactado',
  VISIT = 'visita',
  CLOSED = 'cerrado',
}

export enum LeadIntent {
  BUY = 'buy',
  RENT = 'rent',
  SELL = 'sell',
  VISIT = 'visit',
}